fix: charge move cheats
